+++ ClassDigitizer2Gen.h
@@ -48,7 +48,7 @@ class Digitizer2Gen {
 
     char retValue[256];
 
-    unsigned short serialNumber;
+    unsigned int serialNumber;
     unsigned int CupVer;
     std::string  FPGAType; // look the DigitiParameter.h::PHA::DIG::FirwareType, DPP_PHA, DPP_ZLE, DPP_PSD, DPP_DAW, DPP_OPEN, and Scope
     unsigned int FPGAVer; // for checking copy setting
@@ -92,13 +92,13 @@ class Digitizer2Gen {
     Digitizer2Gen();
     ~Digitizer2Gen();
 
-    unsigned short GetSerialNumber() const {return serialNumber;}
+    unsigned int GetSerialNumber() const {return serialNumber;}
     std::string    GetFPGAType()     const {return FPGAType;}
     std::string    GetModelName()    const {return ModelName;}
     unsigned int   GetFPGAVersion()  const {return FPGAVer;}
     unsigned int   GetCupVer()       const {return CupVer;}
 
-    void  SetDummy(unsigned short sn);
+    void  SetDummy(unsigned int sn);
     bool  IsDummy()     const {return isDummy;}
     bool  IsConnected() const {return isConnected;}
